+++ packages/ant-design/src/__tests__/Description/faas.test.tsx
@@ -17,7 +17,7 @@ describe('Description/faas', () => {
         text: async () => Promise.resolve('{"data":{"test":"value"}}'),
       }) as unknown as Promise<Response>
     })
-    FaasReactClient({ domain: 'test' })
+    FaasReactClient({ domain: '/' })
   })
 
   afterEach(() => {