+++ packages/ant-design/src/__tests__/Table/ajax.test.tsx
@@ -30,7 +30,7 @@ describe('Table/ajax', () => {
           }),
       }) as unknown as Promise<Response>
     })
-    FaasReactClient({ domain: 'test' })
+    FaasReactClient({ domain: '/' })
   })
 
   afterEach(() => {