+++ packages/react/src/client.tsx
@@ -5,10 +5,11 @@ import type {
   useFaasOptions,
 } from './types'
 import type { FaasAction, FaasData, FaasParams } from '@faasjs/types'
-import type { Options, Response, ResponseError } from '@faasjs/browser'
+import type { BaseUrl, Options, Response, ResponseError } from '@faasjs/browser'
 import { FaasBrowserClient } from '@faasjs/browser'
 import { FaasDataWrapper } from './FaasDataWrapper'
 import { useFaas } from './useFaas'
+import { faas } from './faas'
 
 const clients: {
   [key: string]: FaasReactClientInstance
@@ -20,7 +21,7 @@ export type OnError = (
 ) => (res: ResponseError) => Promise<void>
 
 export type FaasReactClientOptions = {
-  domain: string
+  domain: BaseUrl
   options?: Options
   onError?: OnError
 }
@@ -28,14 +29,14 @@ export type FaasReactClientOptions = {
 /**
  * Before use faas, you should initialize a FaasReactClient.
  *
- * @param props.domain {string} The domain of your faas server
+ * @param props.host {string} The domain of your faas server
  * @param props.options {Options} The options of client
  * @returns {FaasReactClientInstance}
  *
  * @example
  * ```ts
  * const client = FaasReactClient({
- *   domain: 'localhost:8080/api'
+ *   domain: 'localhost:8080/api/'
  * })
  * ```
  */
@@ -46,24 +47,13 @@ export function FaasReactClient({
 }: FaasReactClientOptions): FaasReactClientInstance {
   const client = new FaasBrowserClient(domain, options)
 
-  async function faas<PathOrData extends FaasAction>(
-    action: PathOrData | string,
-    params: FaasParams<PathOrData>,
-    options?: Options
-  ): Promise<Response<FaasData<PathOrData>>> {
-    if (onError)
-      return client
-        .action<PathOrData>(action, params, options)
-        .catch(async res => {
-          await onError(action as string, params)(res)
-          return Promise.reject(res)
-        })
-    return client.action(action, params, options)
-  }
-
   const reactClient = {
     id: client.id,
-    faas,
+    faas: async <PathOrData extends FaasAction>(
+      action: PathOrData | string,
+      params: FaasParams<PathOrData>,
+      options?: Options
+    ): Promise<Response<FaasData<PathOrData>>> => faas(action, params, options),
     useFaas: <PathOrData extends FaasAction>(
       action: PathOrData | string,
       defaultParams: FaasParams<PathOrData>,
@@ -72,8 +62,9 @@ export function FaasReactClient({
       useFaas(action, defaultParams, options),
     FaasDataWrapper: <PathOrData extends FaasAction>(
       props: FaasDataWrapperProps<PathOrData>
-    ) => <FaasDataWrapper domain={domain} {...props} />,
+    ) => <FaasDataWrapper baseUrl={domain} {...props} />,
     onError,
+    browserClient: client,
   }
 
   clients[domain] = reactClient
@@ -84,18 +75,18 @@ export function FaasReactClient({
 /**
  * Get FaasReactClient instance
  *
- * @param domain {string} empty string for default domain
+ * @param host {string} empty string for default host
  * @returns {FaasReactClientInstance}
  *
  * @example
  * ```ts
  * getClient()
  * // or
- * getClient('another-domain')
+ * getClient('another-host')
  * ```
  */
-export function getClient(domain?: string): FaasReactClientInstance {
-  const client = clients[domain || Object.keys(clients)[0]]
+export function getClient(host?: string): FaasReactClientInstance {
+  const client = clients[host || Object.keys(clients)[0]]
 
   if (!client) {
     console.warn('FaasReactClient is not initialized manually, use default.')
@@ -106,25 +97,4 @@ export function getClient(domain?: string): FaasReactClientInstance {
   }
 
   return client
-}
-
-/**
- * Request faas server
- *
- * @param action {string} action name
- * @param params {object} action params
- * @returns {Promise<Response<any>>}
- *
- * @example
- * ```ts
- * faas<{ title: string }>('post/get', { id: 1 }).then(res => {
- *   console.log(res.data.title)
- * })
- * ```
- */
-export async function faas<PathOrData extends FaasAction>(
-  action: string | PathOrData,
-  params: FaasParams<PathOrData>
-): Promise<Response<FaasData<PathOrData>>> {
-  return getClient().faas(action, params)
 }